fix(Prefab): ensure combined rotation axes are continuously processed
The movement axes are passed to a moment processor so they are
continually processed each frame to ensure the button press keeps
moving the avatar even when the button is still being held down.

This concept has now also been applied to the rotational combined
actions as they were only picking up immediate changes in the input
which worked fine when the mouse was the input type but meant that
using an xbox controller for rotational input left the movement
jerky because it would only move when the axis value changed.