feat(prefab): add device details about headset and controllers
The new DeviceDetailsRecord has been extended into the
SimulatorNodeRecord that provide information about a simulated node
device and this is now stored on the prefab so it is easily available
at runtime.

The DominantControllerObserver has also been added to the prefab to
make it easier to determine the current dominant controller.